requirements.txt
colourlovers
Pillow